# docs/swagger.yaml
basePath: /
definitions:
...
  main.ErrorResponse:
    properties:
      code:
        type: integer
      error:
        type: string
    type: object
...
  main.User:
    properties:
      avatar_url:
        type: string
      created_at:
        type: integer
      email:
        type: string
      id:
        type: integer
      is_admin:
        description: 添加管理員標記
        type: boolean
      name:
        type: string
      social_id:
        type: string
      social_provider:
        type: string
      updated_at:
        type: integer
    type: object
  TokenResponse:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
      expires_in:
        type: integer
    type: object
host: localhost:8080
info:
  contact: {}
  description: 這是一個買一送一配對網站的API文檔
  title: 買一送一配對網站 API
  version: "1.0"
paths:
...
  /auth/{provider}:
    get:
      consumes:
      - application/json
      description: 開始 Facebook 或 Instagram OAuth 流程
      parameters:
      - description: OAuth 提供者 (facebook 或 instagram)
        in: path
        name: provider
        required: true
        type: string
      produces:
      - application/json
      responses:
        "302":
          description: 重定向到 OAuth 提供者
          schema:
            type: string
        "500":
          description: OAuth 開始失敗
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      summary: 開始 OAuth 流程
      tags:
      - 認證
  /auth/{provider}/callback:
    get:
      consumes:
      - application/json
      description: 處理 OAuth 提供者的回調
      parameters:
      - description: OAuth 提供者 (facebook 或 instagram)
        in: path
        name: provider
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 使用者資訊和 tokens
          schema:
            properties:
              user:
                $ref: '#/definitions/main.User'
              access_token:
                type: string
              refresh_token:
                type: string
              expires_in:
                type: integer
            type: object
        "400":
          description: 無效的提供者
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "500":
          description: OAuth 回調錯誤
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      summary: OAuth 回調處理
      tags:
      - 認證
  /auth/token:
    get:
      consumes:
      - application/json
      description: 將現有的 session 交換為 JWT token
      produces:
      - application/json
      responses:
        "200":
          description: JWT tokens
          schema:
            $ref: '#/definitions/TokenResponse'
        "401":
          description: 未登入
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      summary: 交換 session for JWT token
      tags:
      - 認證
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: 使用 refresh token 獲取新的 access token 和 refresh token
      parameters:
      - description: Refresh token request
        in: body
        name: request
        required: true
        schema:
          properties:
            refresh_token:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 新 tokens
          schema:
            $ref: '#/definitions/TokenResponse'
        "400":
          description: 無效的請求
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "401":
          description: 無效的 refresh token
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      summary: Refresh access token
      tags:
      - 認證
  /logout:
    get:
      consumes:
      - application/json
      description: 登出使用者並清除 session
      produces:
      - application/json
      responses:
        "302":
          description: 重定向到首頁
          schema:
            type: string
        "500":
          description: 登出失敗
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      summary: 處理登出
      tags:
      - 認證
...
  /profile:
    get:
      consumes:
      - application/json
      description: 取得使用者資訊 (支援 session 和 JWT token 認證)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.User'
        "401":
          description: 未登入
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "500":
          description: 無法取得使用者資訊
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: 取得使用者資訊
      tags:
      - 使用者
...
securityDefinitions:
  ApiKeyAuth:
    description: 輸入 'Bearer &lt;JWT token&gt;' 進行認證。可以先透過 Facebook 登入取得 token。
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"